Fix DB reset problem
Prev the db used to reset and the api was recalled and restored
on every server start up. This was fine during development but during
production it doesnt allow for persistance. The problem was solved by
adding middleware to check if todays data is stored and if it is dont
call the api and if it isnt to call the API and store the data.